Fix iteration in lowering.
Lowering was assuming that the node being lowered would never be removed
and was therefore terminating early in cases that it was (e.g. lowering
of an add to an lea on x64). This changes the lowering helpers to return
the lowered node which is then used as the iteration cursor.