Be less conservative about removing dead stores.
The current backend happily removes dead stores with the GTF_ORDER bit
set unless the value being stored is a catch arg. The LIR implementation
was a bit more conservative, refusing to remove dead stores with any of
the GTF_ALL_EFFECT bits set. Change this to match the current
implementation.