Refactor some of the LIR utilities.
The most interesting change here is that LIR::Range is no longer final and has been
split into two types: ReadOnlyRange and Range. The latter is used for subranges; the
former is used for ranges that are not subranges (i.e. blocks and scratch ranges).

LIR::Range has also been made the base type of BasicBlock, thus obviating the need
for the hokey pointer tricks it used to play for polymorphism. Because LIR::Range
is no longer a final class, it is now possible to slice range values. As a result,
viewing a BasicBlock as an LIR::Range now returns a reference to the range rather
than the range itself.

Although some pains have been taken to help guard against the existence of invalid
subranges, it remains possible to create them. For example, consider the following:

    // View the block as a range
    LIR::Range& blockRange = LIR::AsRange(block);

    // Create a range from the first non-phi node in the block to the last node in
    // the block
    LIR::ReadOnlyRange nonPhis = blockRange.NonPhiNodes();

    // Remove the last node from the block
    blockRange.Remove(blockRange.LastNode());

After the removal of the last node in the block, the last node of `nonPhis` is no
longer linked to any of the other nodes in `nonPhis`.

Other, smaller changes include:
- LIR::Range::{Begin,End,EndExclusive} are dead. Long live {FirstNode(),EndNode()}.
- Some code snippets that were calculating the last node of a block inline now use
  BasicBlock::lastNode
- BasicBlock::IsLIR() may now return true for an empty block. Blocks are marked as
  LIR using a bit stolen from `bbNum`. HIR blocks are transitioned to LIR blocks
  as part of rationalize, and all blocks that are created after rationalize are
  automatically marked as LIR blocks.