Fixes for x86 LIR
With this, crossgen of x86 S.P.C.dll is as clean as the non-LIR version. There is a known
assert with checked. With debug, there are a number of crashes with JMP calls that also
occur with the legacy JIT.

There is one big caveat: I added an NYI in LowerArg for long stack arguments. This hits 450 times.
The code is:
```
// TODO(pdg): the code below uses tree order to freely reorder the lo and hi pieces of
// a long-type argument. This needs to be fixed.
NYI("LIR: long arg lowering");
```
This will be fixed next.

The changes:
1. During decomposition, make sure tree costs are set.
2. Fix bug in LIR::Range::rend() if the range is empty.
3. In CheckLIR(), make sure the forward and reverse threadings have no circularities.

Also:
a. Dump MethodHash with COMPLUS_JitFunctionTrace