Introduce LIR::Range::Delete.
This method combines the previous sequence of calls to Remove and
DecRefCnts into a single method. This method also calls
DEBUG_DESTROY_NODE on the deleted nodes.

This change revealed a bug in dead store removal: if a store in a
`gtCallArgs` list was found to be dead (this can with copy
propagation), the removed store was still referred to by the
`gtCallArgs` list. This change fixes that bug by bashing the store
to a NOP if it is marked with the `GTF_LATE_ARG` flag.