Add a dominator-based CSE pass.
This pass imlpements CSE as a depth-first walk of the dominator tree
that finds and replaces available CSEs as it goes. This approach
drastically reduces the amount of time spent in CSE: the dominator
CSE pass is retires about 65% fewer instructions than the dataflow CSE
pass when precompiling `System.Private.CoreLib.dll`, which translates
to about as 1% reduction in overall instructions retired for the same.
Because the pass is cheaper overall, it does not limit the number of
CSEs it attempts to perform. This differs from the dataflow pass, which
limits the number of CSE candidates to 64.

This pass does not currently perform any profitability heuristics
beyond "can I CSE this operator"; instead, it assumes that any available
CSE will be an improvement. This pass may also be somewhat less
detrimental to downstream optimizations, as it is trivially able to
maintain value numbers (and could in the future be updated to create SSA
for CSE temps if desired).

The code size impact of this change is something of a mixed bag: overall
code size regresses by a small amount (mostly due to RA behvaior), but
many methods see a (somtimes significant) code size improvement.
Furthermore, some of the code size regressions may result in
improvements in execution time, as the new pass is more aggressive about
removing small (but redundant) expressions (e.g. heap loads).